+++ moor/lib/src/runtime/query_builder/expressions/algebra.dart
@@ -7,7 +7,8 @@ extension MonoidExpr<DT, ST extends Monoid<DT>> on Expression<DT, ST> {
     assert(other is! Expression<String, StringType>,
         'Used Monoid extension but should have resolved to String extension');
 
-    return _BaseInfixOperator(this, '+', other);
+    return _BaseInfixOperator(this, '+', other,
+        precedence: Precedence.plusMinus);
   }
 }
 
@@ -16,7 +17,8 @@ extension StringMonoidExpr on Expression<String, StringType> {
   /// Performs a string concatenation in sql by appending [other] to `this`.
   Expression<String, StringType> operator +(
       Expression<String, StringType> other) {
-    return _BaseInfixOperator(this, '||', other);
+    return _BaseInfixOperator(this, '||', other,
+        precedence: Precedence.stringConcatenation);
   }
 }
 
@@ -25,7 +27,8 @@ extension ArithmeticExpr<DT, ST extends FullArithmetic<DT>>
     on Expression<DT, ST> {
   /// Performs a subtraction (`this` - [other]) in sql.
   Expression<DT, ST> operator -(Expression<DT, ST> other) {
-    return _BaseInfixOperator(this, '-', other);
+    return _BaseInfixOperator(this, '-', other,
+        precedence: Precedence.plusMinus);
   }
 
   /// Returns the negation of this value.
@@ -35,11 +38,13 @@ extension ArithmeticExpr<DT, ST extends FullArithmetic<DT>>
 
   /// Performs a multiplication (`this` * [other]) in sql.
   Expression<DT, ST> operator *(Expression<DT, ST> other) {
-    return _BaseInfixOperator(this, '*', other);
+    return _BaseInfixOperator(this, '*', other,
+        precedence: Precedence.mulDivide);
   }
 
   /// Performs a division (`this` / [other]) in sql.
   Expression<DT, ST> operator /(Expression<DT, ST> other) {
-    return _BaseInfixOperator(this, '/', other);
+    return _BaseInfixOperator(this, '/', other,
+        precedence: Precedence.mulDivide);
   }
 }