+++ moor/lib/src/runtime/query_builder/expressions/bools.dart
@@ -5,16 +5,18 @@ part of '../query_builder.dart';
 /// This is now deprecated. Instead of `and(a, b)`, use `a & b`.
 @Deprecated('Use the operator on BooleanExpressionOperators instead')
 Expression<bool, BoolType> and(
-        Expression<bool, BoolType> a, Expression<bool, BoolType> b) =>
-    _AndExpression(a, b);
+    Expression<bool, BoolType> a, Expression<bool, BoolType> b) {
+  return a & b;
+}
 
 /// Returns an expression that is true iff [a], [b] or both are true.
 ///
 /// This is now deprecated. Instead of `or(a, b)`, use `a | b`;
 @Deprecated('Use the operator on BooleanExpressionOperators instead')
 Expression<bool, BoolType> or(
-        Expression<bool, BoolType> a, Expression<bool, BoolType> b) =>
-    _OrExpression(a, b);
+    Expression<bool, BoolType> a, Expression<bool, BoolType> b) {
+  return a | b;
+}
 
 /// Returns an expression that is true iff [a] is not true.
 ///
@@ -31,43 +33,26 @@ extension BooleanExpressionOperators on Expression<bool, BoolType> {
 
   /// Returns an expression that is true iff both `this` and [other] are true.
   Expression<bool, BoolType> operator &(Expression<bool, BoolType> other) {
-    return _AndExpression(this, other);
+    return _BaseInfixOperator(this, 'AND', other, precedence: Precedence.and);
   }
 
   /// Returns an expression that is true if `this` or [other] are true.
   Expression<bool, BoolType> operator |(Expression<bool, BoolType> other) {
-    return _OrExpression(this, other);
+    return _BaseInfixOperator(this, 'OR', other, precedence: Precedence.or);
   }
 }
 
-class _AndExpression extends _InfixOperator<bool, BoolType> {
-  @override
-  Expression<bool, BoolType> left, right;
-
-  @override
-  final String operator = 'AND';
-
-  _AndExpression(this.left, this.right);
-}
-
-class _OrExpression extends _InfixOperator<bool, BoolType> {
-  @override
-  Expression<bool, BoolType> left, right;
-
-  @override
-  final String operator = 'OR';
-
-  _OrExpression(this.left, this.right);
-}
-
 class _NotExpression extends Expression<bool, BoolType> {
-  Expression<bool, BoolType> inner;
+  final Expression<bool, BoolType> inner;
 
   _NotExpression(this.inner);
 
   @override
+  Precedence get precedence => Precedence.unary;
+
+  @override
   void writeInto(GenerationContext context) {
     context.buffer.write('NOT ');
-    inner.writeInto(context);
+    writeInner(context, inner);
   }
 }